# core, eth/filters, miner, xeth: Optimised log filtering