@obscuren 2015-10-12
This PR still requires a database update strategy to add all receipt log addresses to the mip mapped blooms.