@obscuren 2015-10-12
```
benchmark              old ns/op      new ns/op     delta
BenchmarkMipmaps-4     3786039581     102561295     -97.29%
```