@obscuren 2015-10-12
@karalabe the full-node PR is going to need some changes for this because we should still be able to search through past receipts. Wherever in the code receipts are being added it's going to require an additional `core.WriteMipmapBloom(db, blockNumber, receipts)` this so that receipts can be included to the mip mapped bloom bins.