@karalabe 2015-10-13
@obscuren Sure, just ping me after this gets merged, or if the other way around, I'll tell you where to insert this (i.e. `BlockChain.InsertReceiptChain`).

Btw, just a question with assigning receipts to block numbers (haven't looked at your code yet). I see you assign it to block numbers. Can you handle cases where there are multiple blocks/receipts with the same number? Currently in the database everything (apart from the canonical mapping) is mapped to hashes, so scenarios like this don't occur. Just trying to make sure this doesn't go misinterpreted :)