@obscuren 2015-10-13
@karalabe nothing gets assigned to anything actually. There are general bloom bins for several levels (0-`1m-1`, 0-`500k-1`, 500k-`1m-1`, 0-`100k-1`, etc). The keys for these are generated as `N / L • L` where `L` is the level in which you are storing.

All of this cares nothing for hashes. There are potentially higher rates of false positives due to chain re-orgs but I think that's quite alright.