@frozeman 2015-10-13
I just tried your branch, it doesn't give me the past logs, even though to returns fast.
Does it map also previous logs, or do i need to re-sync my whole chain to make it work?