@obscuren 2015-10-13
@frozeman yes full resynx until I get the database merge feature in (see first comment)