@obscuren 2015-10-13
Added upgrade strategy and tests for upgrading