@obscuren 2015-10-13
@karalabe PTAL changed the `TempDir`