@drupalnomad 2015-10-15
How does this compare with the cpp-ethereum approach?