@obscuren 2015-10-15
@drupalnomad what do you mean? Which approach