@drupalnomad 2015-10-16
@obscuren Gav said that cpp-ethereum already has indexing for searching logs, but in #1895 you say this would take up too much disk space. Maybe there is just confusion over definition of "indexing"