@karalabe 2015-10-16
Indexing is a fairly expensive and complex operation. You need to maintain a separate "index" database (or dataset at least), which you need to keep in sync with block insertions, chain reorganisations, etc. I don't know exactly what the performance or storage hit would be, but probably non negligible.

However, if you think about it, all indexing would need to do is create a presence list. Is something present or not at a certain block. A full fledged index is much more powerful that that (hence the associated costs). This PR on the other hand uses a different (and imho quite a brilliant idea), of creating bloom filters at various resolutions, that simply state whether a given log has or has not a chance of being present at some section of the chain (e.g. from 100K to 110K). The bloom filters can provide false positives, but will never produce false negatives (i.e. if it says a log isn't present in the interval, no way can it be present). So you can quickly zoom in on sections of the chain that feature a certain log, and then iterate the blocks to find it (which are few by count).