@obscuren 2015-10-16
@drupalnomad I've no idea how C++ is doing it, in fact I wasn't even aware C++ had something like indexing (or some other technique).