@drupalnomad 2015-10-16
Gav said this:
>however for cpp, at least, there's no reason we can't apply the same scalability to lightclient logic and provide very fast topic searches over millions of blocks
>the topic limit is actually 4 (3 indexed parameters in solidity + the indexed event name)