@obscuren 2015-10-16
Ah, the `indexed` is a keyword in `solidity`. The `indexed` keywords indicates the compiler that the argument is something that is to be "**searchable**" through filtering; e.g. `event T(i indexed uint)` and when called, internally in the VM, calls `LOG1(value)`. This `value` then becomes filterable.

If we had an `event T(i uint)` the `i` argument won't be filterable but instead is chucked in the **extra** `data` field.

Logs are triplets in the form of `[address, [topics], data]`. All `indexed` parameters are considered `topics` with a maximum of 4.