@karalabe 2015-10-16
:+1: LGTM now, but note, **I never used logs, so I do not know how to properly live test this**. @frozeman Please check out this current version and play with it and report any issues.